feat(ayako): current_user_id in context + HasFavourited field in BeatmapSet

+++ ayako/middlewares/permission/local.go
@@ -11,7 +11,7 @@ var MustLogin = func(next echo.HandlerFunc) echo.HandlerFunc {
 	return func(c echo.Context) error {
 		data := c.Get("current_user_id")
 
-		_, ok := data.(uint64)
+		_, ok := data.(uint)
 		if !ok {
 			return echo.NewHTTPError(http.StatusUnauthorized, "Invalid token")
 		}
